Calculate propeller thrust and torque correctly
The main challange was to figure out the fact that the thrust and
torque values computed from thrust and torque coefficients had kgf
unit instead of N. The other issue was motor line-to-line resistance
which was originally guessed to be 0.5Ohm. That value was quite wrong.
The value is updated to something more realistic now.